Package References

Package References

ApplyNoise

ReduceNoise

NoiseAPI

ApplyNoise

ApplyNoise module provides several noise prototypes.

ImageNoise.ApplyNoise.AdditiveWhiteGaussianNoise — Type

AdditiveWhiteGaussianNoise <: AbstractImageNoise
AdditiveWhiteGaussianNoise([μ=0.0], σ)

apply white gaussian noise to image

For gray images, it uses the following formula:

out = clamp01.(in .+ σ .* randn(size(in)) .+ μ)

RGB images are treated as 3D-gray images, generic Color3 images will be converted to RGB images first.

Examples

img = testimage("lena_gray_256")
n = AdditiveWhiteGaussianNoise(0.1)
out = apply_noise(img, n)

See also: apply_noise, apply_noise!

References

[1] Wikipedia contributors. (2019, March 8). Additive white Gaussian noise. In Wikipedia, The Free Encyclopedia. Retrieved 14:32, June 9, 2019, from https://en.wikipedia.org/w/index.php?title=AdditivewhiteGaussian_noise&oldid=886818982

ReduceNoise

ReduceNoise module provides several noise reduction algorithms.

ImageNoise.ReduceNoise.NonlocalMean — Type

NonlocalMean(λ [, r_p=2, r_s=2r_p+1])
NonlocalMean(λ, img)

Non-local mean denoising filter for additive white Gaussian noise using the following formula:

\[ y[p] = \sum_{q \in \mathscr{N}(r\_s, p)} w[p, q]x[q]\]

where w[p, q] represents the similarity of two patches centered in p and q, calculated by

\[ w[p, q] = \frac{1}{Z[p]}e^{-\frac{\lVert\mathscr{N}(r\_w, p) - \mathscr{N}(r\_w, q) \rVert^2_{2, a}}{λ^2}}\]

Z[p] is the normalizing constant so that ∑w over q equals to 1.

Arguments

λ::Float64 is the degree of filtering, larger λ produces a smoother result.

r_p::Int64 is the radius of image patch size. By default it's 2.

r_s::Int64 is the radius of search window, large r_s would slow down

the filtering significantly. By default it's 2r_p + 1.

If you pass img to NonlocalMean, it will be used to estimate r_p and r_s.

Examples

img = testimage("lena_color_256")

n = AdditiveWhiteGaussianNoise(0.1)
noisy_img = apply_noise(img, n)

# use default filter arguments
f_denoise = NonlocalMean(0.1)
denoised_img = reduce_noise(noisy_img, f)

# estimate filter arguments with noisy_img
f_denoise = NonlocalMean(0.1, noisy_img)
denoised_img = reduce_noise(noisy_img, f)

See also: reduce_noise, reduce_noise!

References

[1] Buades, A., Coll, B., & Morel, J. M. (2005, June). A non-local algorithm for image denoising. In 2005 IEEE Computer Society Conference on Computer Vision and Pattern Recognition (CVPR'05) (Vol. 2, pp. 60-65). IEEE.

ImageNoise.ReduceNoise._colon — Method

_colon(I:J)

_colon(I, J) works equivelently to I:J, it's used to backward support julia v"1.0".

ImageNoise.ReduceNoise.make_kernel — Method

gaussian-like kernel

NoiseAPI

NoiseAPI is an experimental module on unifying the API of different denoise algorithms. This is used by algorithm develpers, users are not expected to use this module directly.

ImageNoise.NoiseAPI.AbstractImageAlgorithm — Type

AbstractImageAlgorithm

The root of image algorithms type system

ImageNoise.NoiseAPI.AbstractImageFilter — Type

AbstractImageFilter <: AbstractImageAlgorithm

Filters are image algorithms whose input and output are both images

ImageNoise.NoiseAPI.apply_noise — Function

apply_noise([::Type,] img, n::AbstractImageNoise, args...; rng=GLOBAL_RNG, kwargs...)

Add/Apply noise n to image img.

Examples

n = AdditiveWhiteGaussianNoise(0.1)
img = testimage("lena_gray_256")
apply_noise(img, n)

# sometimes we need to pass `rng` keyword argument to
# generate reproducible noise
apply_noise(img, n; rng = MersenneTwister(0))

See also: apply_noise!

ImageNoise.NoiseAPI.apply_noise! — Function

apply_noise!([out,] img, n::AbstractImageNoise, args...; rng=GLOBAL_RNG, kwargs...)

Add/Apply noise n to image img.

If out is specified, it will be changed in place. Otherwise img will be changed in place.

Examples

n = AdditiveWhiteGaussianNoise(0.1)
img = testimage("lena_gray_256")
apply_noise!(img, n)

# sometimes we need to pass `rng` keyword argument to
# generate reproducible noise
apply_noise!(img, n; rng = MersenneTwister(0))

See also: apply_noise

ImageNoise.NoiseAPI.reduce_noise — Function

reduce_noise([::Type,] img, f::AbstractImageDenoiseAlgorithm, args...)

Remove noise of image img using algorithm f.

See also: reduce_noise!

ImageNoise.NoiseAPI.reduce_noise! — Function

reduce_noise!([out,] img, f::AbstractImageDenoiseAlgorithm, args...)

Remove noise of image img using algorithm f.

If out is specified, it will be changed in place. Otherwise img will be changed in place.

See also: reduce_noise

ImageNoise.NoiseAPI.@filter_api — Macro

@filter_api api_name [filter_type=AbstractImageFilter]

This macro it generate two methods:

api_name([::Type,] img, f::filter_type, args...; kwargs...)

api_name!([out,] img, f::filter_type, args...; kwargs...)

For in-place method api_name!, out will be changed after calling the method. When out is not explicitly passed, img will be changed after calling the method.

Info

Any api implementation needs to support a f(out, in, args...) method.

This macro is designed to be used in ImagesAPIs, and not downstream packages

Example:

1. register the api in ImagesAPI.jl

abstract type AbstractImageNoise <: AbstractImageFilter end
@filter_api apply_noise AbstractImageNoise

2. implement the api in ImageNoise.jl

import Main.ImagesAPI: AbstractImageNoise, AbstractImageFilter, apply_noise, apply_noise!

export
    apply_noise, apply_noise!,
    AbstractImageNoise,
    AdditiveWhiteGaussianNoise

struct AdditiveWhiteGaussianNoise{T<:AbstractFloat} <: AbstractImageNoise
    mean::T
    std::T
end

function (noise::AdditiveWhiteGaussianNoise)(out, in::AbstractArray)
    @. out = in + noise.std * randn(eltype(out), size(in)) + noise.mean
end

3. user call the API in a consistent way

using ImageNoise
noise = AdditiveWhiteGaussianNoise(0.0, 0.1)

# simple usage
apply_noise(ones(3,3), noise)

# inplace changing
img = ones(3,3)
apply_noise!(img, noise)

# preallocation output
img = ones(3,3)
out = zeros(3, 3)
apply_noise!(out, img, noise)